64-bit audit: global particle counts and orders (required before the 2B dataset)
Widened every type that holds a GLOBAL particle count or order (int
overflows at 2.147e9; the next dataset is ~2B):
- Particle::order (the global id and the FoF tip value domain)
- BoundingBox n_particles/n_sph/n_dark/n_star
- Reader start_particle (the product n_particles*thisIndex overflows
  before orders are even assigned)
- ThreadStateHolder counters + collectMetaData tuple (sum reaches N);
  Driver consumer + Meta print formats
- Decomposition saved_n_total_particles, splitter thresholds,
  GenericSplitter::goal_rank (ranks in [0,N)), SFC and OCT
  countAssignments count vectors + sum_long reductions (bisection counts
  span up to N; the weighted variant sums partition indices over
  particles)
- DecompArrayMap ctor, Subtree ctor, Writer ctor/prefix counts (+ .ci
  signature updates); deleteParticleOfOrder
- FoFPhase1 flat_order (global orders); uf_parent stays int (PE-local
  indices) with the reasoning recorded
- phaseB/star-emit dedup keys: exact TipPairKey replaces the 32|32
  packing that silently collides (suppressing real edges) above 4.29e9
  particles — same bug class phase 3 fixed in step 4
- fof3 harness N + format strings

Boundaries recorded, not silently crossed: Tipsy header fields are int32
BY FORMAT — TipsyWriter now CkEnforces N <= 2^31-1 with a comment (use
NChilada or split outputs beyond); Kd/Binary/LongestDim decompositions
keep int counts, flagged as not audited (unused by current apps).

Validation: 12-run matrix, 1M b0.2 (333889) and b0.8 (41315)
grid-verified, LAMBS 1M (379884), 8M stats line bit-identical,
fof1/annotate/searchAlgos pass. Correctness ABOVE 2^31 is validated by
type reasoning, not by run — no 2B input fits the laptop.

+++ examples/fof3/FoF3.C
@@ -37,9 +37,9 @@ using namespace paratreet;
     // value Driver::run later passes to traversalFn).
     const BoundingBox& universe = thread_state_holder.ckLocalBranch()->universe;
     double V = universe.box.volume();
-    int N = universe.n_particles;
+    long N = universe.n_particles;
     fof_b = fof_b_factor * std::cbrt(V / (double)N);
-    CkPrintf("FoF linking length b = %g (factor %g, V = %g, N = %d)\n",
+    CkPrintf("FoF linking length b = %g (factor %g, V = %g, N = %ld)\n",
              fof_b, fof_b_factor, V, N);
 
     // Phase 1: register -> phaseA -> phaseB -> merge -> relabel. Tips are
@@ -364,7 +364,7 @@ using namespace paratreet;
     // Phase 1 + upwardPass already ran in preTraversalFn (before loadCache;
     // see the ordering note there). Sanity-check that the linking length
     // computed there matches this universe.
-    int N = universe.n_particles;
+    long N = universe.n_particles;
     double b = fof_b;
     CkEnforce(b > 0.0);
     CkEnforce(std::fabs(b - fof_b_factor *
@@ -389,7 +389,7 @@ using namespace paratreet;
 
     // Self-describing config line: every FOF3STAT block starts with this.
     // pbc = the cubic box period L (0 = open boundaries; design/pbc.md).
-    CkPrintf("FOF3STAT config: pes %d nodes %d N %d b %.12g b_factor %g "
+    CkPrintf("FOF3STAT config: pes %d nodes %d N %ld b %.12g b_factor %g "
              "decomp %s tree %s leafsize %d mode %s pbc %g walk %s iter %d\n",
              CkNumPes(), CkNumNodes(), N, b, fof_b_factor,
              paratreet::asString(conf.decomp_type).c_str(),
@@ -449,7 +449,7 @@ using namespace paratreet;
     // separately by Partition::verifySharedLeaves inside runFoFPhase3[Dist].)
     double tg0 = CkWallTimer();
     if (uf2_mode == UF2Mode::Dist) paratreet::runFoFVerifyEncodedTips(fof);
-    else                           paratreet::runFoFVerifyTips(fof, (long)N);
+    else                           paratreet::runFoFVerifyTips(fof, N);
     double tg1 = CkWallTimer();
     CkPrintf("FOF3STAT time_s: fragcheck %.3f tip_sentinel %.3f\n",
              tc1 - tc0, tg1 - tg0);
@@ -549,7 +549,7 @@ using namespace paratreet;
     // global order of the component's min-order member -- the same canonical
     // representative the serial reference computes -- so compare directly.
     double tv0 = CkWallTimer();
-    CkReductionMsg* msg = gatherRecords(fof, N, "final");
+    CkReductionMsg* msg = gatherRecords(fof, (int)N, "final"); // gated: full verify only runs at N <= kAutoFullMaxN
     double tv1 = CkWallTimer();
     const auto* recs = (const FoFParticleRecord*)msg->getData();
     // Reference partition: exact grid-hash FoF (O(n) at fixed density). At
@@ -558,7 +558,7 @@ using namespace paratreet;
     // the grid runs — the O(n^2) path stays available behind this size
     // threshold).
     std::vector<long> serial_rep;
-    long serial_components = gridReference(recs, N, b, b * b, pbc, serial_rep);
+    long serial_components = gridReference(recs, (int)N, b, b * b, pbc, serial_rep);
     double tv2 = CkWallTimer();
     if (N <= 10000) {
       std::vector<long> n2_rep;
